requirement.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
# Order matters to the pip dependency resolver, so sorting this file
# changes how packages are installed.  New dependencies should be
# added in alphabetical order, however, some dependencies may need to
# be installed in a specific order.
#
# PBR should always appear first
pbr>=1.8
# Griver Core Requirements
Django>=1.9
djangorestframework>=3.3.2
#eventlet>=0.17.4
#httplib2>=0.7.5
#six>=1.9.0
pycrypto>=2.6.1
markdown>=2.6.6
pygments

